Move linux-only user namespace and mount code into its own file
